fixed deleting answers security

+++ app/Http/Controllers/AppController.php
@@ -78,16 +78,16 @@ class AppController extends Controller
     public function delete_answer($answer_id)
     {
         $answer = Answer::find($answer_id)->find($answer_id);
-//        dd($answer);
-        $answer->delete();
-
+        if(Auth::user() && Auth::user()->id == $answer->responder_id)
+            $answer->delete();
         return redirect(url('answers/'.$answer->question_id));
     }
 
     public function delete_question($question_id)
     {
         $question = Question::find($question_id);
-        $question->delete();
+        if(Auth::user() && Auth::user()->id == $question->asker_id)
+            $question->delete();
     }
 }
